requirements.txt
asgiref==3.9.1
bcrypt==4.3.0
Django==5.2.5
django-filter==25.2
djangorestframework==3.16.1
mysqlclient==2.2.7
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
